Split common funtions to library
